Fixes #809: LibTest\core\double tests migrated to null safety.

+++ LibTest/core/double/floorToDouble_A01_t04.dart
@@ -37,8 +37,8 @@ main() {
   check(1234567.0E20);
   check(1234567.0E40);
   check(1234567.0E80);
-  check(Math.pow(2.0, 52));
-  check(Math.pow(2.0, 53));
-  check(Math.pow(2.0, 520));
+  check(Math.pow(2.0, 52) as double);
+  check(Math.pow(2.0, 53) as double);
+  check(Math.pow(2.0, 520) as double);
   check(1.0E308);
 }
