+++ LibTest/core/double/toStringAsFixed_A03_t01.dart
@@ -20,12 +20,12 @@ import "../../../Utils/expect.dart";
 import "dart:math";
 
 main() {
-  double x1 = pow(10.0, 21);
+  double x1 = pow(10.0, 21) as double;
   Expect.equals("1e+21", x1.toStringAsFixed(0));
   Expect.equals("1e+21", x1.toStringAsFixed(10));
   Expect.equals("1e+21", x1.toStringAsFixed(20));
 
-  double y1 = pow(10.0, 20);
+  double y1 = pow(10.0, 20) as double;
   Expect.equals("100000000000000000000", y1.toStringAsFixed(0));
   Expect.equals("100000000000000000000.0000000000", y1.toStringAsFixed(10));
   Expect.equals(
